src/data/usecases/db-save-survey-result.ts
import { SurveyResultModel } from '@/domain/models'
import { SaveSurveyResult, SaveSurveyResultParams } from '@/domain/usecases'
import { SaveSurveyResultRepository, LoadSurveyResultRepository } from '@/data/protocols'

export class DbSaveSurveyResult implements SaveSurveyResult {
  constructor (
    private readonly saveSurveyResultRepository: SaveSurveyResultRepository,
    private readonly loadSurveyResultRepository: LoadSurveyResultRepository

  ) { }

  async save (surveyResultData: SaveSurveyResultParams): Promise<SurveyResultModel> {
    await this.saveSurveyResultRepository.save(surveyResultData)
    const loadedSurveyResult =
    await
    this.loadSurveyResultRepository
      .loadBySurveyId(surveyResultData.surveyId)

    return loadedSurveyResult
  }
}
